+++ neuro_san/__init__.py
@@ -31,6 +31,12 @@ _DEPRECATION_REDIRECT = DeprecationRedirect(
     __name__,
     # A map from old class name to new class name for compatibility
     {
+        "neuro_san.internals.authorization.interfaces.abstract_authorizer.AbstractAuthorizer":
+            "neuro_san.service.authorization.interfaces.abstract_authorizer.AbstractAuthorizer":
+        "neuro_san.internals.authorization.interfaces.authorizer.Authorizer":
+            "neuro_san.service.authorization.interfaces.authorizer.Authorizer":
+        "neuro_san.internals.authorization.openfga.open_fga_authorizer.OpenFgaAuthorizer":
+            "neuro_san.service.authorization.openfga.open_fga_authorizer.OpenFgaAuthorizer":
         "neuro_san.internals.messages.chat_message_type.ChatMessageType":
             "neuro_san.message.types.chat_message_type.ChatMessageType",
         "neuro_san.internals.messages.origination.Origination":
