+++ neuro_san/service/authorization/null/always_yes_authorizer.py
@@ -18,38 +18,14 @@ from typing import Any
 from typing import Dict
 from typing import List
 
-from logging import getLogger
-from logging import Logger
-
-from leaf_common.resolution.resolver import Resolver
-
-from neuro_san.internals.authorization.interfaces.authorizer import Authorizer
+from neuro_san.service.authorization.interfaces.abstract_authorizer import AbstractAuthorizer
 
 
-class AbstractAuthorizer(Authorizer):
+class AlwaysYesAuthorizer(AbstractAuthorizer):
     """
-    Partial Authorizer implementation that allows for late-binding via a resolver.
+    Implementation of the Authorizer interface that lets all requests through.
+    This gives us the behavior we have had since the beginning of Neuro SAN.
     """
-
-    def __init__(self):
-        """
-        Constructor.
-        """
-        self.resolver: Resolver = Resolver()
-        self.logger: Logger = getLogger(self.__class__.__name__)
-
-    async def __aenter__(self) -> Authorizer:
-        """
-        Opens a scoped session with this Authorizer.
-        """
-        return self
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb) -> bool:
-        """
-        Closes a scoped session with this Authorizer.
-        :return: True to suppress exception. False or None to propagate exception.
-        """
-        # Do nothing
 
     async def authorize(self, actor: Dict[str, Any], action: str, resource: Dict[str, Any]) -> bool:
         """
@@ -71,7 +47,8 @@ class AbstractAuthorizer(Authorizer):
         :return: True if the actor is allowed to take the requested action on the resource.
                  False otherwise.
         """
-        raise NotImplementedError
+        # By default, anyone can do anything
+        return True
 
     async def grant(self, actor: Dict[str, Any], relation: str, resource: Dict[str, Any]) -> bool:
         """
@@ -92,7 +69,6 @@ class AbstractAuthorizer(Authorizer):
                         }
         :return: True if the grant succeeded, False if the grant already existed.
         """
-        # Having this interface is more for completeness than fulfilling Neuro SAN server functionality
         return False
 
     async def revoke(self, actor: Dict[str, Any], relation: str, resource: Dict[str, Any]) -> bool:
@@ -114,7 +90,6 @@ class AbstractAuthorizer(Authorizer):
                         }
         :return: True if the revoke succeeded, False if the revoke already existed.
         """
-        # Having this interface is more for completeness than fulfilling Neuro SAN server functionality
         return False
 
     async def list(self, actor: Dict[str, Any], relation: str, resource: Dict[str, Any]) -> List[str]:
@@ -142,33 +117,5 @@ class AbstractAuthorizer(Authorizer):
                  An empty return list implies that the actor has access to no objects
                  of the given resource type.
         """
-        raise NotImplementedError
-
-    async def query(self, actor: Dict[str, Any], relation: str, resource: Dict[str, Any]) -> List[str]:
-        """
-        Instead of a boolean answer from authorize() above, this method gives a list
-        of resources of the given resource type (in the dict) that the actor has the
-        *direct* given relation to.  This does not take authorization policy graphs
-        into account.
-
-        :param actor: The actor dictionary with the keys "type" and "id" identifying what
-                      will be permitted.  Most often this is of the form:
-                        {
-                            "type": "User",
-                            "id": "<username>"
-                        }
-        :param relation: The relation for which the user will be permitted.
-                     Most often this is one of the strings from the Role enum.
-
-        :param resource: The resource dictionary with the keys "type" and "id" identifying
-                      just what is to be authorized for use.  For instance:
-                        {
-                            "type": "AgentNetwork",
-                            "id": "hello_world"
-                        }
-        :return: A list of relations (which can be None or empty) that the actor
-                has the given relation with.
-        """
         # Return None indicating some other mechanism should be used
-        # Having this interface is more for completeness than fulfilling Neuro SAN server functionality
         return None
